Migrate tests for AddActivity from Cucumber to android-testing
Turns out that the setup in LauncherActivityTest does not need
to be that complicated, using an EmptyComposeTestRule to interact
with the compose nodes and an ActivityScenario to control the
activity works just fine. No need to register the activity somehow
with the ComposeTestRule. It just needs to exist as a Rule to
set up the Compose component registry stuff.

But since the plan is to get cucumber running again, I'm leaving
the LauncherActivityTest as is.