Fix "stuck" topbar in launcher on first run
More precisely: when not all setup is complete, and so there's long and
thus on most devices scrollable content in the launcher screen, and
user scrolls down, then fixes the intent receiver situation, then
returns to the launcher screen, the setup cards disappear, scrolling is
not necessary anymore and the top bar stays disappeard and won't come
back (because there's no scrolling anymore).

Not so anymore.